add edit attendance capability, refactor schema
